disk free space disabled hosters

+++ system/libs/GFS/gfs.php
@@ -1345,14 +1345,17 @@ class GFS_Package_installer extends GFS {
 		$tempfolder = ROOT . CACHE_DIRECTORY . "/" . basename($this->file);
 		
 		
-		// check for disk-quote
-		$free = (disk_free_space("/") > disk_free_space(ROOT)) ? disk_free_space(ROOT) : disk_free_space("/");
-		define("GOMA_FREE_SPACE", $free);
-		if($free / 1024 / 1024 < 5) {
-			// free space
-			FileSystem::delete($tempfolder);
-			header("HTTP/1.1 500 Server Error");
-			die(file_get_contents(ROOT . "system/templates/framework/disc_quota_exceeded.html"));
+		$f = @disk_free_space("/");
+		if($f !== null && $f !== "") {
+			// check for disk-quote
+			$free = (disk_free_space("/") > disk_free_space(ROOT)) ? disk_free_space(ROOT) : disk_free_space("/");
+			define("GOMA_FREE_SPACE", $free);
+			if($free / 1024 / 1024 < 5) {
+				// free space
+				FileSystem::delete($tempfolder);
+				header("HTTP/1.1 500 Server Error");
+				die(file_get_contents(ROOT . "system/templates/framework/disc_quota_exceeded.html"));
+			}
 		}
 		
 		// write files
@@ -1700,14 +1703,17 @@ class GFS_Package_Creator extends GFS {
 			$this->fileIndex = $index;
 		}
 		
-		// check for disk-quote
-		$free = (disk_free_space("/") > disk_free_space(ROOT)) ? disk_free_space(ROOT) : disk_free_space("/");
-		define("GOMA_FREE_SPACE", $free);
-		if($free / 1024 / 1024 < 20) {
-			// free space
-			@unlink($this->file);
-			header("HTTP/1.1 500 Server Error");
-			die(file_get_contents(ROOT . "system/templates/framework/disc_quota_exceeded.html"));
+		$f = @disk_free_space("/");
+		if($f !== null && $f !== "") {
+			// check for disk-quote
+			$free = (disk_free_space("/") > disk_free_space(ROOT)) ? disk_free_space(ROOT) : disk_free_space("/");
+			define("GOMA_FREE_SPACE", $free);
+			if($free / 1024 / 1024 < 20) {
+				// free space
+				@unlink($this->file);
+				header("HTTP/1.1 500 Server Error");
+				die(file_get_contents(ROOT . "system/templates/framework/disc_quota_exceeded.html"));
+			}
 		}
 		
 		// Adding files...
